exp: move apply tests into their own module
